requirements.txt
tensorflow==2.0.0
streamlit
numpy
pillow
h5py==2.10.0
opencv-contrib-python-headless